backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Open Email Agent Backend with Composio Integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "composio_openai_agents==0.8.0",
    "openai-agents",
    "fastapi",
    "uvicorn",
    "pydantic[email]",
    "pydantic-settings",
    "python-dotenv",
    "openai",
    "composio",
    "langchain>=0.3.27",
    "langgraph>=0.6.8",
    "langchain-openai>=0.3.33",
    "langchain-mcp-adapters>=0.1.10",
    "composio-langchain>=0.11.1",
    "langchain-anthropic>=1.3.4",
    "langchain-google-genai>=4.2.1",
]
